feat: реализовать регистрацию, добавить валидацию

+++ controllers/cards.js
@@ -1,19 +1,25 @@
 import CardModel from '../models/Card';
 
-export const getCards = async (req, res) => {
+import BadRequestError from '../errors/bad-request-error';
+
+import Forbidden from '../errors/forbidden';
+
+import NotFound from '../errors/not-found';
+
+export const getCards = async (req, res, next) => {
   try {
     const cards = await CardModel.find({});
     if (cards) {
       res.status(200).send(cards);
     } else {
-      res.status(404).send({ message: 'Карточки не найдены' });
+      next(new NotFound('Карточки не найдены'));
     }
   } catch (err) {
-    res.status(500).send({ message: `Произошла ошибка: ${err.name} текст ошибки: ${err.message}` });
+    next(err);
   }
 };
 
-export const createCard = async (req, res) => {
+export const createCard = async (req, res, next) => {
   try {
     const { name, link } = req.body;
     const id = req.user._id;
@@ -21,30 +27,39 @@ export const createCard = async (req, res) => {
     res.status(201).send(newCard);
   } catch (err) {
     if (err.name === 'ValidationError') {
-      res.status(400).send({ message: 'Некорректные данные для создания карточки' });
+      next(new BadRequestError('Некорректные данные для создания карточки'));
     } else {
-      res.status(500).send({ message: `Произошла ошибка: ${err.name} текст ошибки: ${err.message}` });
+      next(err);
     }
   }
 };
 
-export const deleteCardById = async (req, res) => {
-  await CardModel.findByIdAndRemove(req.params.cardId)
+export const deleteCardById = async (req, res, next) => {
+  await CardModel.findById(req.params.cardId)
     .orFail(new Error('NotFound'))
-    .then((responce) => res.status(200).send({ message: `Карточка ${req.params.cardId} удалена ${responce}` }))
-    .catch((err) => {
-      /* console.log(`ВОТ ТУТ!!! ${err}`); */
-      if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные для удаления карточки.' });
-      } else if (err.message === 'NotFound') {
-        res.status(404).send({ message: 'Карточка с указанным id не найдена' });
+    .then((card) => {
+      if (req.user._id === card.owner.toString()) {
+        CardModel.findByIdAndRemove(req.params.cardId)
+          .then((item) => {
+            res.status(200).send({ message: `Карточка ${req.params.cardId} удалена ${item}` });
+          })
+          .catch((err) => next(err));
       } else {
-        res.status(500).send({ message: `Произошла ошибка: ${err.name} текст ошибки: ${err.message}` });
+        next(new Forbidden('Нельзя удалять чужую карточку'));
+      }
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        next(new BadRequestError('Переданы некорректные данные для удаления карточки'));
+      } else if (err.message === 'NotFound') {
+        next(new NotFound('Карточка с указанным id не найдена'));
+      } else {
+        next(err);
       }
     });
 };
 
-export const setLikeByCardId = async (req, res) => {
+export const setLikeByCardId = async (req, res, next) => {
   await CardModel.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
@@ -53,18 +68,17 @@ export const setLikeByCardId = async (req, res) => {
     .orFail(new Error('NotFound'))
     .then((responce) => res.send({ data: responce }))
     .catch((err) => {
-      /* console.log(`ВОТ ТУТ!!! ${err}`); */
       if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные для постановки/снятии лайка.' });
+        next(new BadRequestError('Переданы некорректные данные для постановки/снятии лайка'));
       } else if (err.message === 'NotFound') {
-        res.status(404).send({ message: 'Карточка с указанным id не найдена' });
+        next(new NotFound('Карточка с указанным id не найдена'));
       } else {
-        res.status(500).send({ message: `Произошла ошибка: ${err.name} текст ошибки: ${err.message}` });
+        next(err);
       }
     });
 };
 
-export const unsetLikeByCardId = async (req, res) => {
+export const unsetLikeByCardId = async (req, res, next) => {
   await CardModel.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
@@ -74,13 +88,12 @@ export const unsetLikeByCardId = async (req, res) => {
     .orFail(new Error('NotFound'))
     .then((responce) => res.send({ data: responce }))
     .catch((err) => {
-      /* console.log(`ВОТ ТУТ!!! ${err}`); */
       if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные для постановки/снятии лайка.' });
+        next(new BadRequestError('Переданы некорректные данные для постановки/снятии лайка'));
       } else if (err.message === 'NotFound') {
-        res.status(404).send({ message: 'Карточка с указанным id не найдена' });
+        next(new NotFound('Карточка с указанным id не найдена'));
       } else {
-        res.status(500).send({ message: `Произошла ошибка: ${err.name} текст ошибки: ${err.message}` });
+        next(err);
       }
     });
 };
